+++ modules/consensus/tendermint/primitives/src/lib.rs
@@ -1,0 +1,16 @@
+pub mod verifier;
+pub use cometbft::{
+	block::{Header, Height, signed_header::SignedHeader},
+	chain::Id as ChainId,
+	hash::Hash,
+	time::Time,
+	validator::{Info as Validator, Set as ValidatorSet},
+};
+
+pub use verifier::{
+	ConsensusProof, TrustedState, UpdatedTrustedState, VerificationError, VerificationOptions,
+};
+
+pub mod prover;
+
+pub use prover::ProverError;
