+++ modules/consensus/tendermint/prover/Cargo.toml
@@ -21,6 +21,10 @@ cometbft-proto = { git = "https://github.com/cometbft/cometbft-rs", package = "c
 prost = "0.13"
 sha3 = "0.10"
 tracing = "0.1.40"
+tendermint-primitives = { path = "../primitives" }
+
+
+[dev-dependencies]
 tendermint-verifier = { path = "../verifier" }
 
 [dependencies.polkadot-sdk]
