+++ modules/consensus/tendermint/prover/src/prover.rs
@@ -1,5 +1,5 @@
-use crate::{client::Client, error::ProverError};
-use tendermint_verifier::{ConsensusProof, TrustedState};
+use crate::client::Client;
+use tendermint_primitives::{ConsensusProof, ProverError, TrustedState};
 
 /// Main function to prove a header update
 /// This constructs a consensus proof that can be verified by the verifier
